scanner/requirements.txt
semgrep==1.70.0
pyyaml==6.0.1
requests==2.32.3
